+++ KafkaConsumerRetry/Services/IPartitionMessageManager.cs
@@ -7,13 +7,11 @@ namespace KafkaConsumerRetry.Services;
 public interface IPartitionMessageManager {
     void QueueConsumeResult<TResultHandler>(ConsumeResult<byte[], byte[]> consumeResult) where TResultHandler : IConsumerResultHandler;
 
-    void HandleLostPartitions(IConsumer<byte[], byte[]> consumer, ConsumerConfig consumerConfig,
-        List<TopicPartitionOffset> list);
+    void HandleLostPartitions(IConsumer<byte[], byte[]> consumer, List<TopicPartitionOffset> list);
 
     void HandleAssignedPartitions(IConsumer<byte[], byte[]> consumer, ConsumerConfig consumerConfig,
         List<TopicPartition> list,
         TopicNames topicNames, ProducerConfig producerConfig);
 
-    void HandleRevokedPartitions(IConsumer<byte[], byte[]> consumer, ConsumerConfig consumerConfig,
-        List<TopicPartitionOffset> list);
+    void HandleRevokedPartitions(IConsumer<byte[], byte[]> consumer, List<TopicPartitionOffset> list);
 }