+++ TestConsole/GoBoomException.cs
@@ -1,0 +1,3 @@
+﻿namespace TestConsole;
+
+internal class GoBoomException : Exception { }